requirements.txt
python-telegram-bot==20.7
faster-whisper==1.0.3
httpx==0.27.0
aiohttp==3.9.5
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1
python-dotenv==1.0.1
